Adding GreenDao to project and first test for Repository class + some implementation.

+++ app/src/main/java/com/wildeastcoders/pantroid/model/usecase/RetrievePantryItemUsecase.java
@@ -7,5 +7,5 @@ import com.wildeastcoders.pantroid.model.PantryItem;
  */
 
 public interface RetrievePantryItemUsecase extends Usecase<PantryItem> {
-    void init(int itemId);
+    void init(Long itemId);
 }
